requirements.txt
fastapi
uvicorn[standard]
pydantic-ai
gradio
python-dotenv
requests
httpx
nest_asyncio